--- lib/constants.dart
@@ -1,1 +1,0 @@
-String apiUrl = "https://rickandmortyapi.com/graphql";
