--- lib/data/character.graphql
@@ -1,14 +1,0 @@
-query GetCharacter($id: ID!) {
-  character(id: $id) {
-    id 
-    status
-    type
-    image
-    origin {
-      name
-    }
-    episode {
-      episode
-    }
-  }
-}
