--- lib/data/episode.graphql
@@ -1,12 +1,0 @@
-query GetEpisode ($id: ID!) {
-  episode(id: $id) {
-    id
-    name
-    air_date
-    episode
-    characters {
-      name
-      image
-    }
-  }
-}
