--- lib/data/episodes.graphql
@@ -1,16 +1,0 @@
-query GetEpisodes($page: Int)  {
-  episodes(page: $page) {
-    info {
-      count
-      pages
-      next
-    }
-    results {
-      id
-      name
-      air_date
-      episode
-      created
-    }
-  }
-}
