--- lib/data/schema.graphql
@@ -1,95 +1,0 @@
-directive @cacheControl(
-  maxAge: Int
-  scope: CacheControlScope
-) on FIELD_DEFINITION | OBJECT | INTERFACE
-enum CacheControlScope {
-  PUBLIC
-  PRIVATE
-}
-
-type Character {
-  id: ID
-  name: String
-  status: String
-  species: String
-  type: String
-  gender: String
-  origin: Location
-  location: Location
-  image: String
-  episode: [Episode]
-  created: String
-}
-
-type Characters {
-  info: Info
-  results: [Character]
-}
-
-type Episode {
-  id: ID
-  name: String
-  air_date: String
-  episode: String
-  characters: [Character]
-  created: String
-}
-
-type Episodes {
-  info: Info
-  results: [Episode]
-}
-
-input FilterCharacter {
-  name: String
-  status: String
-  species: String
-  type: String
-  gender: String
-}
-
-input FilterEpisode {
-  name: String
-  episode: String
-}
-
-input FilterLocation {
-  name: String
-  type: String
-  dimension: String
-}
-
-type Info {
-  count: Int
-  pages: Int
-  next: Int
-  prev: Int
-}
-
-type Location {
-  id: ID
-  name: String
-  type: String
-  dimension: String
-  residents: [Character]
-  created: String
-}
-
-type Locations {
-  info: Info
-  results: [Location]
-}
-
-type Query {
-  character(id: ID!): Character
-  characters(page: Int, filter: FilterCharacter): Characters
-  charactersByIds(ids: [ID!]!): [Character]
-  location(id: ID!): Location
-  locations(page: Int, filter: FilterLocation): Locations
-  locationsByIds(ids: [ID!]!): [Location]
-  episode(id: ID!): Episode
-  episodes(page: Int, filter: FilterEpisode): Episodes
-  episodesByIds(ids: [ID!]!): [Episode]
-}
-
-scalar Upload
