+++ lib/queries.dart
@@ -1,0 +1,21 @@
+const String allEpisodesQuery = '''
+    query GetEpisodes (\$page: Int) {
+      episodes (page: \$page) {
+        results {
+          id
+          name
+          episode
+        }
+      }
+    }
+''';
+
+const String singleEpisodeQuery= ''' 
+  query GetEpisode (\$id: ID!) {
+      episode (id: \$id) {
+        characters {
+          name
+      }
+    }
+  }
+''';
